修复 ActiveRingChart 组件 精度丢失问题

+++ src/components/activeRingChart/index.js
@@ -101,7 +101,7 @@ const ActiveRingChart = ({ config = {}, className, style }) => {
 
     const sum = value.reduce((all, v) => all + v, 0)
 
-    const percent = parseInt((value[activeIndex] / sum) * 100) || 0
+    const percent = parseFloat((value[activeIndex] / sum) * 100) || 0
 
     return {
       content: '{nt}%',
